Rename of tests following C# implementation
